delete command simplified, captcha compatibility

+++ src/main/java/pl/suseu/bfactions/command/cmds/DeleteCommandExecutor.java
@@ -2,6 +2,8 @@ package pl.suseu.bfactions.command.cmds;
 
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.player.AsyncPlayerChatEvent;
 import pl.rynbou.langapi3.LangAPI;
 import pl.suseu.bfactions.BFactions;
 import pl.suseu.bfactions.base.field.FieldRepository;
@@ -12,8 +14,11 @@ import pl.suseu.bfactions.base.user.User;
 import pl.suseu.bfactions.base.user.UserRepository;
 import pl.suseu.bfactions.command.BCommand;
 import pl.suseu.bfactions.command.BCommandExecutor;
+import pl.suseu.eventwaiter.EventWaiter;
 
 import java.util.List;
+import java.util.Random;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class DeleteCommandExecutor implements BCommandExecutor {
 
@@ -23,6 +28,7 @@ public class DeleteCommandExecutor implements BCommandExecutor {
     private final GuildRepository guildRepository;
     private final RegionRepository regionRepository;
     private final FieldRepository fieldRepository;
+    private final EventWaiter eventWaiter;
 
     public DeleteCommandExecutor(BFactions plugin) {
         this.plugin = plugin;
@@ -31,6 +37,7 @@ public class DeleteCommandExecutor implements BCommandExecutor {
         this.guildRepository = plugin.getGuildRepository();
         this.regionRepository = plugin.getRegionRepository();
         this.fieldRepository = plugin.getFieldRepository();
+        this.eventWaiter = plugin.getEventWaiter();
     }
 
     @Override
@@ -47,8 +54,27 @@ public class DeleteCommandExecutor implements BCommandExecutor {
             lang.sendMessage("you-do-not-own-any-guild", player);
             return;
         }
-        guild.delete();
-        player.getInventory().addItem(this.plugin.getItemRepository().getItem("beacon", false));
-        lang.sendMessage("guild-deleted", player);
+
+        if (guild.getDeleteCode() != -1) {
+            return;
+        }
+
+        AtomicInteger number = new AtomicInteger(new Random().nextInt(8999) + 1000);
+        guild.setDeleteCode(number.get());
+
+        this.lang.sendMessage("confirm-guild-deletion", player, "%number%", "" + number.get());
+        this.eventWaiter.waitForEvent(AsyncPlayerChatEvent.class, EventPriority.NORMAL,
+                ev -> ev.getPlayer().equals(player) && ev.getMessage().equals("" + number.get()),
+                ev -> {
+                    ev.setCancelled(true);
+                    this.plugin.getServer().getScheduler().runTask(this.plugin, () -> {
+                        player.getInventory().addItem(this.plugin.getItemRepository().getItem("beacon", false));
+                        guild.delete();
+                    });
+                    this.lang.sendMessage("guild-deleted", player);
+                }, 20 * 15, () -> {
+                    this.lang.sendMessage("confirm-guild-deletion-listener-timeout", player);
+                    guild.setDeleteCode(-1);
+                });
     }
 }
